Fix bleeding colors into prompt et al
Chore's color subsystem includes de-duplication efforts to avoid
redundantly printing a color escape value.  This code treats foreground
and background colors as completely independent.  This is correct
behavior for all colors except `Default` which is tied between the two;
setting either foreground or background to `Default` sets _both_.

The robust fix here would be to adjust the system to recognize
`Default`'s nature: if a `Default` request comes in for either
foreground or background, it should be applied if either foreground or
background are non-`Default`.

However, a simpler and adequate solution for the time being is to just
set the foreground `Default` as a closing color rather than the
background's, as background is currently only used for priority which is
almost always followed by other default-backgrounded text.  Thus any
concerns around bleeding colors will almost always be foreground colors.